+++ main.py
@@ -1,6 +1,6 @@
 import json
 import yaml
-from data_processing import vocabulary, data_loader
+from data_processing import vocabulary, graph_dataset
 
 import torch.utils.data.dataset
 
@@ -12,7 +12,7 @@ mode = 'test'
 config = yaml.safe_load(open(config_path))
 vocab = vocabulary.Vocabulary(vocab_path='vocab.txt')
 
-dataset = data_loader.GraphDataset(data_path, vocab, config, mode, debug=False)
+dataset = graph_dataset.GraphDataset(data_path, vocab, config, mode, debug=False)
 
 for sample in dataset:
     tokens, edges, error_loc, repair_targets, repair_candidates = sample
